[MOL-3757][WK] Separate input fn into new component and reverted type
